HW2/step4_ge_validation_data/pyproject.toml
[tool.poetry]
name = "ml-validation-project"
version = "0.1.0"
description = "ML model validation pipeline"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
scikit-learn = "^1.0"
pandas = "^1.0"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
black = "^23.0"
ruff = "^0.1.0"
isort = "^5.12"
pre-commit = "^3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
